pyproject.toml
[build-system]
requires = ["hatchling", "maturin>=1.7,<2"]
build-backend = "hatchling.build"

[project]
name = "buvis-gems"
version = "0.1.0"
description = "BUVIS Python toolkit and CLI tools"
authors = [{ name = "Tomas Bouska", email = "tomas@buvis.net" }]
license = "MIT"
readme = "README.md"
requires-python = ">=3.12,<4.0"
dependencies = [
  "click>=8.1.0,<9",
  "inflection>=0.5.1,<1",
  "pexpect>=4.9.0,<5",
  "pydantic>=2.0,<3",
  "pydantic-settings>=2.0,<3",
  "pyyaml>=6.0.3,<7",
  "rich>=14.2.0,<15",
  "tzlocal>=5.3.1,<6",
  "pywin32>=311,<400; sys_platform == 'win32'",
  "xattr>=1.3.0,<2; sys_platform != 'win32'",
]

[project.optional-dependencies]
bim = ["jira>=3.10.5,<4", "textual>=3,<4", "fpdf2>=2.8,<3"]
hello-world = ["pyfiglet>=1.0.4,<2"]
muc = ["ffmpeg-python>=0.2.0,<1"]
pinger = ["ping3>=5.1.5,<6"]
readerctl = ["requests>=2.32.5,<3"]
bim-web = ["fastapi>=0.115,<1", "uvicorn[standard]>=0.34,<1", "watchfiles>=1.0,<2"]
ml = ["torch>=2.9.1,<3", "transformers>=5.1,<6", "rake-nltk>=1.0.6,<2"]
all = ["buvis-gems[bim,bim-web,hello-world,muc,pinger,readerctl,ml]"]

[project.scripts]
bim = "bim.cli:cli"
dot = "dot.cli:cli"
fctracker = "fctracker.cli:cli"
hello-world = "hello_world.cli:cli"
muc = "muc.cli:cli"
outlookctl = "outlookctl.cli:cli"
pinger = "pinger.cli:cli"
readerctl = "readerctl.cli:cli"
zseq = "zseq.cli:cli"

[project.urls]
Repository = "https://github.com/buvis/gems"
Documentation = "https://buvis.github.io/gems/"

[dependency-groups]
dev = ["maturin>=1.7,<2", "pip-audit>=2.7.3,<3", "pre-commit>=4.5.1,<5", "ruff>=0.6,<1"]
test = ["pytest>=9.0.2,<10", "pytest-mock>=3.15.1,<4", "pytest-cov>=7,<8", "mypy>=1.15,<2", "types-PyYAML>=6,<7", "types-requests>=2,<3", "types-fpdf2>=2,<3"]
docs = ["sphinx>=9,<10", "sphinx-rtd-theme>=3.0.2,<4"]

[tool.hatch.build.targets.wheel]
packages = [
  "src/lib/buvis",
  "src/tools/bim",
  "src/tools/dot",
  "src/tools/fctracker",
  "src/tools/hello_world",
  "src/tools/muc",
  "src/tools/outlookctl",
  "src/tools/pinger",
  "src/tools/readerctl",
  "src/tools/zseq",
]

[tool.hatch.build.targets.wheel.hooks.custom]

[tool.pytest.ini_options]
testpaths = ["tests"]
strict_markers = true
addopts = ["--import-mode=importlib"]

[tool.coverage.run]
source = ["src/lib", "src/tools"]
branch = true

[tool.mypy]
python_version = "3.12"
strict = true
warn_return_any = true
warn_unused_configs = true
mypy_path = "src/lib:src/tools"
namespace_packages = true
explicit_package_bases = true

[[tool.mypy.overrides]]
module = ["torch.*", "transformers.*", "jira.*", "pexpect.*", "rake_nltk.*", "xattr.*", "inflection.*", "nltk.*", "win32com.*", "ffmpeg.*", "ping3.*", "pyfiglet.*", "buvis.pybase.zettel._core"]
ignore_missing_imports = true

[tool.ruff]
target-version = "py312"
line-length = 120

[tool.ruff.lint]
select = ["E", "W", "F", "I", "S", "ANN", "ARG", "SLF", "PLR", "UP", "RUF"]
ignore = ["ANN401", "S105", "S106", "S108", "S603", "S607", "PLR0911", "PLR0912", "PLR0913", "PLR0915", "PLR2004", "RUF012", "UP046", "UP047"]

[tool.ruff.lint.per-file-ignores]
"tests/**/*.py" = ["S101", "ANN", "ARG", "SLF001", "PLR2004", "RUF043", "RUF059", "UP007"]
"src/tools/bim/commands/serve/_actions.py" = ["ARG001"]
"src/tools/bim/commands/serve/_app.py" = ["ARG001"]
"src/tools/bim/commands/serve/_routes.py" = ["ARG001"]
